Kill the started process group in onDestroy
As it stands now, if the user presses the Back button without closing
her shell, it will keep running unattached to any tty.  Over time, this
leads to a buildup of shells, each inaccessible to the user.

Fix this by doing what xterm and other terminals do -- send SIGHUP to
the process group of the started process when quitting.