Add a service to prevent Android from killing our process
By default, an Android process is liable to be killed whenever it's
not in the foreground.  When this happens to us, it leaves the user
unable to resume her terminal session, which is still running.

By starting a service in our process (even one which, like this one,
does nothing at all) and marking it as a foreground service, we prevent
the system from killing our process except in out-of-memory situations.