Add menu options to take WakeLock and WifiLock
It can be very useful to prevent the device from sleeping or from
turning off the wifi radio -- but in a general terminal application it's
difficult to guess when this is wanted.  Consider the following:

(1) Nothing is happening.  We don't want to hold either a WakeLock or
WifiLock.
(2) The user is logged into a remote system via SSH on wifi.  We don't
need to hold a WakeLock, but should take a WifiLock to prevent the SSH
session from being disconnected.
(3) The user is doing some long-running computation (locally on the
device) in the shell.  We should take a WakeLock, but not a WifiLock.
(4) The user is doing apt-get dist-upgrade in her Debian chroot on the
device.  We should hold both a WakeLock and a WifiLock.

Therefore, assume the user knows best, and add menu options to take/drop
a WakeLock and a WifiLock.  Note that this requires adding WAKE_LOCK to
the set of permissions we use.