Change shared library name (again) to make sure our copy takes precedence on CyanogenMod
Commit 28d47b8... ("Fix java.lang.UnsatisfiedLinkError error.") changed
the name of the shared library from libandroidterm to libandroidterm2,
to avoid a potential library conflict with API/ABI incompatible versions
being shipped in some custom firmwares.  Unfortunately, CyanogenMod also
uses the name libandroidterm2:

https://github.com/CyanogenMod/android_packages_apps_AndroidTerm/commit/7d0de89f0ed5231a7d7eb3ca7917accb83c8eaee

Change the name again, this time to something that should hopefully be
unique, which should fix the problems that CyanogenMod users have been
reporting in the Android Market comments.  (A big scary warning to
firmware distributors that the shared library ABI must not be modified
without consulting upstream might be appropriate somewhere ...)

Signed-off-by: Jack Palevich <jackpal@google.com>