Throw away TranscriptScreen backing store on session finish
Android's InputMethodManager seems to hold a reference to a finished
activity object for a while after the activity is destroyed.  This
keeps everything the activity holds a reference to in memory, which in
our case includes a fairly significant chunk of memory: the backing
store to the TranscriptScreen.  Throw this away when it's not needed, to
reduce the memory impact of this behavior.