Avoid accessing a finished TranscriptScreen when resuming the terminal activity
updatePrefs() may ask each terminal session in the list to update its
size, which requires accessing the TranscriptScreen.  Therefore, we need
to update the list of running terminal sessions before calling
updatePrefs(), to preclude the possibility of accessing a finished
TranscriptScreen.

Also, if a window was closed in the WindowList activity, and then the
user selected a specific window, the window number passed to us in
onActivityResult() refers to the window list *after* taking window
closures into account.  We therefore need to defer switching to the
user's selected window until after the list of running sessions is
updated (onActivityResult() is called before onResume(), so it can't be
done there); this avoids calling onResume() on a stale EmulatorView,
which results in an attempt to access a finished TranscriptScreen.