Add a hack to remember the cursor position across screen size changes
As it stands, the cursor ends up immediately after the last
non-whitespace character in the transcript after a screen resize, which
is very frequently incorrect (the last character in the output is
often a space or newline, and it's possible the cursor has been moved,
as during command line editing).

Since lines could be re-wrapped during the screen size update, there's
no easy way to calculate where the cursor should end up.  Therefore, we
use a hack: since ASCII ESC (\033) is non-printable and should never be
part of the transcript, we mark the cursor location with an ESC before
resizing, and restore the cursor to the location where the ESC ends up
afterwards (taking care to save and restore the real character at the
cursor position correctly).

Signed-off-by: Jack Palevich <jackpal@google.com>