Preserve colors and formatting across screen size changes
Currently, getTranscriptText() has no option for preserving
color/formatting information, so this information is lost when resizing
the screen.  Fix this by adding a variant of getTranscriptText() which
accepts a StringBuilder in which color information will be stored, and
using it during resize operations.

To save memory, this implementation only preserves color information for
regions of the screen which contain text, so blank areas of the screen
with colored backgrounds will still revert to the default background
color.

(Why StringBuilder?  Post-Unicode, we cannot store information in the
upper bits of a char, as we would have before.  An ArrayList<Byte> would
be the most natural fit, but the need to box the values creates serious
memory overhead.  StringBuilder handles dynamic sizing of the array for
us and uses less memory per cell than an ArrayList.)

Signed-off-by: Jack Palevich <jackpal@google.com>