Implement fast resize for cases where width remains constant
A resize operation currently requires getting the entire transcript
text, throwing away the backing store, constructing a new one, and then
writing the transcript to the new store.  Particularly when the
transcript is large, this is a slow and inefficient operation.

As it turns out, it's really quite easy to change the number of screen
rows for a particular backing store, which allows much faster resize
operations in the case where the number of columns does not change.
This is particularly useful when using soft keyboards, where the number
of rows will change depending on whether the keyboard is displayed, but
the number of columns doesn't.

Signed-off-by: Jack Palevich <jackpal@google.com>