Display a message in the EmulatorView when the shell exits
Currently, if the shell exits and the option to close windows on shell
exit isn't activated, the user gets no visual cue that the shell has
exited -- only an unresponsive session.  Instead, take a page from the
Mac OS X Terminal's book and display a message that the terminal session
has finished.

Signed-off-by: Jack Palevich <jackpal@google.com>