Correct color rendition for color schemes other than white on black
Applications expect to get black when they ask for a black background,
independent of what the default background color actually is; similarly,
they expect white when they ask for a white foreground, independent of
the default foreground color.  We, however, currently substitute the
default background color for black and the default foreground color for
white in our palette, which leads to incorrect color when the color
scheme is anything other than white on black.

A full solution to this problem probably requires us to distinguish
"default color" from the colors in the palette, which would require
additional storage for color information.  As a band-aid fix, though,
keep track of which colors in the palette are the closest match to the
default colors, and substitute those instead of always substituting for
black and white.

Signed-off-by: Jack Palevich <jackpal@google.com>