Make Exec.createSubprocess() handle environment variables and unlimited args
Add the ability to modify the subprocess's environment by passing in an
array of environment variable strings suitable for passing to putenv()
("VAR=value").  While we're at it, lift the restriction on the number of
arguments passed to the subprocess by using an array instead of a fixed
number of String arguments to hold command line arguments.

This change breaks the ABI for the JNI library, so bump the library ABI
version (to 3) to avoid any potential problems with external copies of
the library.

Signed-off-by: Jack Palevich <jackpal@google.com>