Set TERM in process environment and provide preference to choose value
In ordinary Unix (non-Android) environments, the behavior of programs
that use advanced terminal capabilities depends on the value of TERM in
the environment.  Currently, we don't set TERM at all, which leads most
of those programs to assume a dumb terminal and behave accordingly.

Instead, set TERM in the started process's environment, and provide the
user a preference to choose the value of TERM.  The current default of
"vt100" is the safest, as it's understood on every system since near the
beginning of time, but does not expose capabilities such as colors or
Home/End keys.  The other options provided are "screen", which is the
best match for our currently implemented capabilities and reasonably
widespread, and "linux".

Signed-off-by: Jack Palevich <jackpal@google.com>