Take a high-performance WifiLock on platforms where it's available
A WIFI_MODE_FULL WifiLock does not prevent Android from enabling
power-saving measures which increase latency, decrease throughput, and
break multicast, all of which are frequently important to programs
running in the terminal which need the WifiLock.  Instead, on API 12 or
later, take a WIFI_MODE_FULL_HIGH_PERF WifiLock, which keeps the wifi
radio in the same state as when the system is awake (if the driver
supports it).

Signed-off-by: Jack Palevich <jackpal@google.com>