Let external GestureListener intercept GestureDetector events in EmulatorView
For Honeycomb and ICS, we want to be able to make a screen tap bring up
the Action Bar in fullscreen mode.  We could in principle do this by
letting EmulatorView hold an ActionBar object, but it seems more
sensible to add hooks into EmulatorView to allow an external
GestureListener, and have the activity use those to bring up the Action
Bar instead.  This also allows us to move the existing
swipe-to-change-window code into the activity, eliminating the need to
hold a TermViewFlipper in the EmulatorView.

Signed-off-by: Jack Palevich <jackpal@google.com>