Add support for autohiding action bar
Offer a preference on Honeycomb and later to allow the action bar to be
hidden.  When hidden, the action bar can be displayed by tapping on the
upper half of the EmulatorView or by pressing the Menu key (if
available).  Tapping on the bottom half of the EmulatorView brings up
the soft keyboard (unless a hardware keyboard is attached, in which case
it also brings up the action bar).

Signed-off-by: Jack Palevich <jackpal@google.com>