Make toasts legible.
Center them, rather than the default, which is to show them near the bottom
of the screen. The old position would show the toasts on top of the
soft keyboard, where they were hard to see.

Also make the reset message stay up longer since it's more text to read.