Refactor session and widget code into separate library project
We'd like to make the terminal emulator and view reusable by others, so
separate out this code (primarily TermSession, TerminalEmulator,
TranscriptScreen, EmulatorView, and associated classes) into a library
usable by other projects.  Users of the new library are expected to
instantiate (or subclass) TermSession, connect it to the emulation
client via an InputStream and OutputStream, and then attach it to an
EmulatorView, which will display the terminal screen and scrollback.

In the app, we subclass both TermSession (ShellTermSession) and
EmulatorView (TermView) to consolidate the setting and updating of
preferences in one place.  Since tty-related handling stays in the app
(there's no guarantee library users will be using a tty), we also
override various methods of TermSession to handle our tty functionality.