Update LayoutParams and redo layout on EmulatorView resize
At the moment, when we resize an EmulatorView, we're only resizing the
contents and the area on which we're drawing, not the view itself.  This
is fine in most cases (any blank space is being covered by the soft
keyboard), but on ICS, when the status bar is disabled and the action
bar is always showing, rotating from portrait to landscape with the soft
keyboard showing appears to result in an EmulatorView large enough to
push the action bar off screen.

Work around this by resizing the EmulatorView itself through adjustment
of its LayoutParams, not just the contents.

Signed-off-by: Jack Palevich <jackpal@google.com>