Delay allocation of the transcript until after we know the terminal's size
During the modularization for by the multisession patches, the lazy
initialization which used to ensure that the transcript wasn't allocated
until its final size was known was lost.  This results in the transcript
always being allocated at least twice for a newly created session
(unless the final width of the window is exactly 80 columns).

To prevent this, delay allocation of the transcript in TermSession
until after the size is set with a call to updateSize().  This requires
some changes in the EmulatorView's initialization code to ensure that
the TranscriptScreen and TerminalEmulator objects aren't used before
they're allocated.

Signed-off-by: Jack Palevich <jackpal@google.com>