Handle Android 3.0+ full keyboards correctly.
Our toggling behavior was confusing people, especially with respect
to the CTRL key on Android 3.0+ full physical keyboards.

We now never toggle when we receive the physical ctrl key. The physical
control key is only present on Android 3.0+ full keyboards.

We now leave it to the system to handle physical caps lock. Caps lock
is only present on Android 3.0+ full keyboards.

We don't toggle shift or alt on keyboards that report that
they are MODIFIER_BEHAVIOR_CHORDED. (i.e. Android 3.0+ full keyboards.)