Make Hangul jamo medial vowels and final consonants combining
While modern Korean is typically written using precomposed syllable
blocks, Unicode also provides individual conjoining Hangul syllables
(jamo) which are supposed to combine to form complete syllable blocks.
Since at least Android 4.1, the Android text rendering services handle
conjoining jamo correctly -- but in order for a syllable block to be
rendered correctly, we need to pass its components to the text renderer
as a unit.

The easiest way to achieve this is to treat medial vowels and final
consonants as combining characters; that way, UnicodeTranscript will
automatically store them in the same column along with the initial
consonant, and all our text-processing code will treat the syllable
block as a unit.  This behavior is strictly incorrect (isolated
individual medials and finals shouldn't behave as combining characters),
but it's the easiest way to get conjoining jamo working in the common
cases.

On platforms where we don't trust Android to render conjoining jamo
correctly, we instead treat medials and finals as regular spacing
characters of width 2.  This is also strictly incorrect (the Unicode
character database says medials and finals are East Asian
neutral/narrow), but it's the best match for the way Android renders
the individual jamo.