Remove inaccessible directories from the user's PATH
Go through the default PATH and remove directories without the execute
bit set (meaning we can't access files in that directory).  Requires
File.canExecute() from API 9 or later.  This will prevent the misleading
"permission denied" errors which come up when you try to run a
nonexistent command; instead, the correct "not found" error will appear.

Add a preference to control the behavior (defaulting to true) so that
people who insist on always running as root can turn this off.