Allow external packages to modify PATH by responding to broadcasts
One of the most frequently-requested features in Android Terminal
Emulator is an easy way to install additional command-line tools beyond
the limited set that Android offers by default.

We accomplish this goal here by sending out ordered broadcasts when
we're first started (in onCreate()), asking for external packages to
provided additional directories to be added to the user's PATH.  This
allows Market packages to make commands available to ATE users by
implementing a BroadcastReceiver which will respond to the broadcast
with the directory it's installed its tools in.

We send out two broadcasts with two different actions:

* jackpal.androidterm.broadcast.APPEND_TO_PATH requests directories to
  be appended to the user's PATH.  No special permissions are required
  to receive this broadcast.
* jackpal.androidterm.broadcast.PREPEND_TO_PATH requests directories to
  be prepended to the user's PATH.  This is intended to allow BusyBox
  installers to override system-installed commands, and is therefore
  protected by a new permission,
  jackpal.androidterm.permission.PREPEND_TO_PATH.

In addition, two new preferences are added to allow the user to disable
all PATH extensions, or just PATH prepending.