Make library internals package-private
Many of the classes, fields, and methods in the library don't need to be
exposed to users.  Make them package-private or private.

To accommodate the changing of TerminalEmulator to package-private, we
need to add a method to TermSession to allow writing directly to the
TerminalEmulator input.  Do this and convert the existing user of this
functionality.