EmulatorView library: use default scheme if setColorScheme() not called
The documentation (and common sense) suggests that it shouldn't be
necessary to call setColorScheme() on the EmulatorView and TermSession
before using them, but this currently causes crashes.  Fix this by
introducing a default color scheme (corresponding to white on black,
with the color values as defined in the default BaseTextRenderer color
arrays).