TermSession: add processInput() method to allow subclasses to modify input data
It's useful for derived classes to be able to modify the input stream
before passing it on to the TerminalEmulator (e.g. to interpret protocol
commands).  At the moment, this requires that they wrap the real
InputStream with a filtering component; instead, give them a method to
override, which should generally be more convenient.