TermSession: Add a write() method which functions on byte arrays
This will be more convenient for many use cases than the existing
methods, which operate on Strings and individual Unicode code points.
Also, by modifying the existing methods (which encode their input to
UTF-8 bytes) to use this one for pushing data to the writer thread, we
give subclasses a single method to override if they need to modify the
output stream.