Store window title in TermSession and hook up display in UI
The window title is logically a property of the terminal session, so
store it as a field in the TermSession object, which allows us to just
do session.getTitle() (modulo minor complications) in TermViewFlipper
and WindowListAdapter when we want the session title.  (In practice, we
also add a getTitle() to ShellTermSession which takes a default title
argument to use in case of the title being null or empty, and use that
instead.)

We do still need a notification interface for when the title changes,
but we no longer need to pass around the title, so use the existing
UpdateCallback interface for that.  Also, extend SessionList to offer a
pass-through title change notification interface, which makes the
implementation in WindowListAdapter easier.