Poll for window size change unconditionally on API < 8
Automatic window resize appears to be broken on API < 8 when the status
bar is visible and the device is in landscape mode.

Strictly speaking, we could disable polling in portrait mode with the
status bar visible, but detecting orientation is a bit annoying and the
once-per-second polling has no effects other than wasting a little CPU
time in that case.  Instead, unconditionally poll for window size
changes on older platforms.