Fix crash when resizing transcript containing FullUnicodeLines
Commit 7135a50... ("Fix rotation of colored text.") introduced checking
of color and formatting attributes as well as text when determining the
last printing character on a line.  However, the naive approach to doing
this results in an ArrayIndexOutOfBoundsException when the text in
question contains non-BMP characters or combining characters, because in
those cases, the char array will be longer than the color array.

Fix this by using the correct index into the color array, the screen
column of the character.