Guard against custom implementations of InputConnection
This should fix AbstractMethodError in EmulatorView$3.getSelectionStart(),
which has been occasionally seen in the Exception logs.

We extend a BaseInputConnection rather than implement an InputConnection
in order to avoid an AbstractMethodError in the case where we are running
on a custom ROM that has added a method to InputConnection. (We've seen
this occur a few timesin the exception logs.)

Using a BaseInputConnection rather than an InputConnection is a little bit
error prone because it means we won't get a compiler warning or error when
a new method has been added to InputConnection.