Fix F1-F4 escape sequences.
Add a comment describing what escape sequences we're trying to match.

This patch is noisy because the previous version of this file didn't use
project-standard whitespace. The whitespace has been fixed.

Thanks to Yi Zhang <yiz.cheapwine@gmail.com> for reporting the bug and
providing a patch that fixed it.