Fix inserting wide char into column previously occupied by normal-width char
Since an East Asian wide character takes two screen columns, if we're
inserting one into a spot previously occupied by a normal-width
character, we need to overwrite the contents of the column to the right
of the position into which the character is inserted.  We do this
shifting the contents of the rest of the line left in the array, then
updating our idea of the storage used -- or at least that's what's
supposed to happen.

There are at least two problems with the current implementation:

(1) findStartOfColumn() is used in a few places to get the spot in the
array where a column starts, but this relies on the offset array, which
is potentially out of date at this point.  Fix this by adding the shift
(which will be used to update the offset array later) to the value
whenever we use findStartOfColumn().
(2) When truncating the line by reducing the storage used count, we
currently effectively reduce the count by the size of the last column in
the line, which is not necessarily the same as the size of the
overwritten column.  We should instead adjust by the size of the
just-overwritten column (which is readily available to us at this
point).

Fixes #145, and potentially a class of other difficult-to-reproduce bugs
involving East Asian wide character support.