Fix rendering of non-BMP characters at start or end of text run
We currently discover a non-BMP character at the start or end of a
text run only after we process both halves of the surrogate pair
representing the character individually, resulting in off-by-one errors
in calculating the portion of the char array containing the run to be
drawn (we erroneously omit any high surrogate at the start of a run, and
erroneously include any high surrogate immediately following the end of
a run).

Fix this by processing both halves of the surrogate pair immediately
when we see the high surrogate.