Remove initial default command.
I think for most users of Android Terminal emulator the default initial
command (which adds /data/local/bin to the path) was not useful,
because the directory /data/local/bin does not exist by default.

The /data/local directory is special because its permissions are set
so that it can be written to by the Android shell even on non-rooted
phones.

Therefore, it is possible for users of non-rooted phones to create the
/data/local/bin directory, and then put executables into that dir,
and then add it to their path, and then execute them as shell commands.

But, in reality, most people who would go to the trouble of installing
command-line applications tend to root their phone and set their
phone up more like a normal Linux system.

So there were very few people (possibly none) for which the default
initial command string was useful.