Update translations for recent changes to Android Terminal Emulator.
These files were all from the Gingerbread branch of CyanogenMod's version
of Android-Terminal-Emulator. They didn't have the latest strings, which
is a problem for Array items, because it means that some entries were
missing from the UI. (For example you couldn't pick large font sizes in
these languages.)

I added all the missing strings. In some cases I copied from a similar
language. In other cases I was able to guess at a translation based on the
text that was already in the file. And in other cases I just used the
original English thext.