Avoid NPE in EmulatorView.onResume in Telnet sample.
That sample calls EmulatorView.attachSession asynchronously, which means
that mKeyListener can be null at the point in time when onResume is called.

For good luck add the same null pointer check to onPause.

I am not 100% happy with this. There are many other places in
EmulatorView where mKeyListener is called without the null pointer check.

But it is a simple band-aid fix that avoids the problem we are seeing.

Good enough for now.

A long term fix probably involves rethinking whether we allow attachSession
to be called asynchronously, and if we do, how to bullet-proof the code.

(At the very least attachSession should be required to be called
on the UI thread.)