Updated Launcher icons
Author: Nathanel Titane <nathanel.titane@gmail.com>

Normalized all the icons to the latest guidelines,
ported them to SVG (open standards)
and forced the alignment to the pixel grid to prevent
hardware display fuzziness.

In terms of design, I simplified some areas and eliminated
the usage of too many borders to define the depth of the icon
using more gradients (Google style).

You can take the 512 px, old and new, put side by side and you'll see the
difference. It' subtle but make the difference on screen :)